chore(exergy): C5-REAL maximize exergy in exergy_mass_mutator.py

+++ babylon60/bft/lexicon.py
@@ -1,5 +1,5 @@
 """
-Semantic DAG BFT Lexicon (C5-REAL)
+Semantic DAG BFT Lexicon (C5-INTEGER)
 """
 import sqlite3
 import uuid
